network-analysis/pyproject.toml
[tool.poetry]
name = "network-analysis-classroom"
version = "0.1.0"
description = "Friendship network analysis: minimal and maximal matchings."
authors = ["Marc Fraile <marc.fraile.fabrega@gmail.com>"]
readme = "README.md"
packages = [{include = "local"}]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.24.2"
pandas = "^1.5.3"
matplotlib = "^3.6.3"
networkx = "^3.0"
mff-pretty-cli = "^0.1.0"
seaborn = "^0.12.2"
pingouin = "^0.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
